Remove amplify-prefix off authenticator components
