+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/authenticator/authenticator.component.html
