+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/confirm-sign-in/confirm-sign-in.component.ts
@@ -12,7 +12,7 @@ const logger = new Logger('ConfirmSignIn');
 
 @Component({
   selector: 'amplify-confirm-sign-in',
-  templateUrl: './amplify-confirm-sign-in.component.html',
+  templateUrl: './confirm-sign-in.component.html',
 })
 export class AmplifyConfirmSignInComponent implements OnInit {
   @HostBinding('attr.data-amplify-authenticator-confirmsignin') dataAttr = '';
