+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/confirm-sign-up/confirm-sign-up.component.ts
@@ -3,7 +3,7 @@ import { AuthenticatorService } from '../../../../services/authenticator.service
 import { translate } from '@aws-amplify/ui';
 @Component({
   selector: 'amplify-confirm-sign-up',
-  templateUrl: './amplify-confirm-sign-up.component.html',
+  templateUrl: './confirm-sign-up.component.html',
 })
 export class AmplifyConfirmSignUpComponent {
   @Input() headerText = translate('Confirm Sign Up');
