+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/federated-sign-in-button/federated-sign-in-button.component.ts
@@ -4,7 +4,7 @@ import { AuthenticatorService } from '../../../../services/authenticator.service
 
 @Component({
   selector: 'amplify-federated-sign-in-button',
-  templateUrl: './amplify-federated-sign-in-button.component.html',
+  templateUrl: './federated-sign-in-button.component.html',
 })
 export class AmplifyFederatedSignInButtonComponent {
   @Input() provider: FederatedIdentityProviders;
