+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/force-new-password/force-new-password.component.html
