+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/index.ts
@@ -1,14 +1,14 @@
-export { AmplifyAuthenticatorComponent } from './amplify-authenticator/amplify-authenticator.component';
-export { AmplifyConfirmSignInComponent } from './amplify-confirm-sign-in/amplify-confirm-sign-in.component';
-export { AmplifyConfirmSignUpComponent } from './amplify-confirm-sign-up/amplify-confirm-sign-up.component';
-export { AmplifyFederatedSignInButtonComponent } from './amplify-federated-sign-in-button/amplify-federated-sign-in-button.component';
-export { AmplifyFederatedSignInComponent } from './amplify-federated-sign-in/amplify-federated-sign-in.component';
-export { AmplifyForceNewPasswordComponent } from './amplify-force-new-password/amplify-force-new-password.component';
-export { AmplifyResetPasswordComponent } from './amplify-reset-password/amplify-reset-password.component';
-export { AmplifySetupTotpComponent } from './amplify-setup-totp/amplify-setup-totp.component';
-export { AmplifySignInComponent } from './amplify-sign-in/amplify-sign-in.component';
-export { AmplifySignUpComponent } from './amplify-sign-up/amplify-sign-up.component';
-export { AmplifySignUpFormFieldsComponent } from './amplify-sign-up/sign-up-form-fields/sign-up-form-fields.component';
-export { AmplifyVerifyUserComponent } from './amplify-verify-user/amplify-verify-user.component';
+export { AmplifyAuthenticatorComponent } from './authenticator/authenticator.component';
+export { AmplifyConfirmSignInComponent } from './confirm-sign-in/confirm-sign-in.component';
+export { AmplifyConfirmSignUpComponent } from './confirm-sign-up/confirm-sign-up.component';
+export { AmplifyFederatedSignInButtonComponent } from './federated-sign-in-button/federated-sign-in-button.component';
+export { AmplifyFederatedSignInComponent } from './federated-sign-in/federated-sign-in.component';
+export { AmplifyForceNewPasswordComponent } from './force-new-password/force-new-password.component';
+export { AmplifyResetPasswordComponent } from './reset-password/reset-password.component';
+export { AmplifySetupTotpComponent } from './setup-totp/setup-totp.component';
+export { AmplifySignInComponent } from './sign-in/sign-in.component';
+export { AmplifySignUpComponent } from './sign-up/sign-up.component';
+export { AmplifySignUpFormFieldsComponent } from './sign-up/sign-up-form-fields/sign-up-form-fields.component';
+export { AmplifyVerifyUserComponent } from './verify-user/verify-user.component';
 export { ConfirmResetPasswordComponent } from './confirm-reset-password/amplify-confirm-reset-password.component';
 export { ConfirmVerifyUserComponent } from './confirm-verify-user/amplify-confirm-verify-user.component';
