+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/reset-password/reset-password.component.html
