+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/setup-totp/setup-totp.component.html
