+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/sign-in/sign-in.component.html
