+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/sign-in/sign-in.component.ts
@@ -9,7 +9,7 @@ import { translate } from '@aws-amplify/ui';
 
 @Component({
   selector: 'amplify-sign-in',
-  templateUrl: './amplify-sign-in.component.html',
+  templateUrl: './sign-in.component.html',
   encapsulation: ViewEncapsulation.None,
 })
 export class AmplifySignInComponent {
