+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/sign-up/sign-up.component.html
